fix(e2e): resolve conftest import collision between unit and e2e tests
