+++ tests/e2e/conftest.py
@@ -6,67 +6,18 @@ a working renderdoc installation (GPU marker applied automatically).
 
 from __future__ import annotations
 
-import json
 import os
-import subprocess
 import uuid
 from collections.abc import Generator
 from pathlib import Path
-from typing import Any
 
 import pytest
-
-FIXTURES_DIR = Path(__file__).parent.parent / "fixtures"
-VKCUBE = FIXTURES_DIR / "vkcube.rdc"
-HELLO_TRIANGLE = FIXTURES_DIR / "hello_triangle.rdc"
-DYNAMIC_RENDERING = FIXTURES_DIR / "dynamic_rendering.rdc"
-OIT_DEPTH_PEELING = FIXTURES_DIR / "oit_depth_peeling.rdc"
-VKCUBE_VALIDATION = FIXTURES_DIR / "vkcube_validation.rdc"
-
-
-# ---------------------------------------------------------------------------
-# Helpers
-# ---------------------------------------------------------------------------
-
-
-def rdc(
-    *args: str,
-    session: str = "e2e_default",
-    timeout: int = 30,
-) -> subprocess.CompletedProcess[str]:
-    """Run ``uv run rdc`` as a subprocess and return the result."""
-    cmd = ["uv", "run", "rdc", "--session", session, *args]
-    return subprocess.run(
-        cmd,
-        capture_output=True,
-        text=True,
-        timeout=timeout,
-    )
-
-
-def rdc_ok(*args: str, session: str = "e2e_default", timeout: int = 30) -> str:
-    """Run rdc, assert exit 0, return stdout."""
-    r = rdc(*args, session=session, timeout=timeout)
-    assert r.returncode == 0, f"rdc {' '.join(args)} failed:\n{r.stderr}"
-    return r.stdout
-
-
-def rdc_json(*args: str, session: str = "e2e_default", timeout: int = 30) -> Any:
-    """Run rdc with --json, assert exit 0, return parsed JSON."""
-    out = rdc_ok(*args, "--json", session=session, timeout=timeout)
-    return json.loads(out)
-
-
-def rdc_fail(
-    *args: str, session: str = "e2e_default", exit_code: int = 1, timeout: int = 30
-) -> str:
-    """Run rdc, assert expected non-zero exit, return combined output."""
-    r = rdc(*args, session=session, timeout=timeout)
-    assert r.returncode == exit_code, (
-        f"Expected exit {exit_code}, got {r.returncode}\nstdout: {r.stdout}\nstderr: {r.stderr}"
-    )
-    return r.stdout + r.stderr
-
+from e2e_helpers import (
+    DYNAMIC_RENDERING,
+    OIT_DEPTH_PEELING,
+    VKCUBE,
+    rdc,
+)
 
 # ---------------------------------------------------------------------------
 # Session fixtures
