+++ tests/e2e/test_assert.py
@@ -10,7 +10,7 @@ from __future__ import annotations
 from pathlib import Path
 
 import pytest
-from conftest import rdc, rdc_fail, rdc_ok
+from e2e_helpers import rdc, rdc_fail, rdc_ok
 
 pytestmark = pytest.mark.gpu
