+++ tests/e2e/test_export.py
@@ -9,7 +9,7 @@ from __future__ import annotations
 from pathlib import Path
 
 import pytest
-from conftest import rdc_fail, rdc_ok
+from e2e_helpers import rdc_fail, rdc_ok
 
 pytestmark = pytest.mark.gpu
