+++ tests/e2e/test_query.py
@@ -10,7 +10,7 @@ from __future__ import annotations
 import re
 
 import pytest
-from conftest import rdc, rdc_fail, rdc_ok
+from e2e_helpers import rdc, rdc_fail, rdc_ok
 
 pytestmark = pytest.mark.gpu
