+++ tests/e2e/test_shader_edit.py
@@ -13,7 +13,7 @@ import json
 from pathlib import Path
 
 import pytest
-from conftest import rdc, rdc_ok
+from e2e_helpers import rdc, rdc_ok
 
 pytestmark = pytest.mark.gpu
