+++ tests/e2e/test_vfs.py
@@ -12,7 +12,7 @@ from __future__ import annotations
 from pathlib import Path
 
 import pytest
-from conftest import rdc_fail, rdc_ok
+from e2e_helpers import rdc_fail, rdc_ok
 
 pytestmark = pytest.mark.gpu
